Wrangle monit service into shape.
There's been some issues between restart_monit and the service monit.
Moonshine would use the init.d script for managing the service, and
relies on it having a 'status' to know if it's running. Without it, it
always tries to start it. Which would fail if it doesn't exist.

This replies the init.d with /lib/init/upstart-job, and uses upstart to
manage it instead. This makes things a lot saner.